MOVIES.json relocated
